requirements.txt
kivymd2
dotenv
langchain_community
praw
langchain_google_genai
langgraph
geopy
kivy_garden
garden install webview